pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vidpost"
version = "0.1.0"
description = "Multi-platform video posting CLI tool"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "rich>=13.0",
    "apscheduler>=3.10",
    "faster-whisper>=1.0",
    "ffmpeg-python>=0.2",
    "google-api-python-client>=2.100",
    "google-auth-oauthlib>=1.2",
    "requests-oauthlib>=1.3",
]

[project.scripts]
vidpost = "vidpost.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/vidpost"]
